Refactor lookupAsn
